requirements.txt
PyYAML~=6.0
PySide6~=6.2.0
sortedcontainers~=2.4.0
pymediainfo~=5.1.0
numpy~=1.21.3
crontab~=0.23.0
